Move all section related files to 'sections' directory
