+++ lib/components/sections/achievement_section.dart
@@ -3,10 +3,10 @@ import 'package:brunstadtv_app/graphql/queries/achievements.graphql.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import 'achievement_dialog.dart';
-import 'achievement_list_item.dart';
-import 'horizontal_slider.dart';
-import '../graphql/queries/page.graphql.dart';
+import '../achievement_dialog.dart';
+import '../achievement_list_item.dart';
+import '../horizontal_slider.dart';
+import '../../graphql/queries/page.graphql.dart';
 
 class AchievementSection extends StatelessWidget {
   final Fragment$Section$$AchievementSection data;
