+++ lib/components/sections/icon_grid_section.dart
@@ -1,13 +1,13 @@
-import 'package:brunstadtv_app/components/section_item_click_wrapper.dart';
+import 'package:brunstadtv_app/components/sections/section_item_click_wrapper.dart';
 import 'package:brunstadtv_app/models/analytics/sections.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import '../graphql/queries/page.graphql.dart';
-import '../graphql/schema/pages.graphql.dart';
-import 'custom_grid_view.dart';
-import 'category_button.dart';
+import '../../graphql/queries/page.graphql.dart';
+import '../../graphql/schema/pages.graphql.dart';
+import '../custom_grid_view.dart';
+import '../category_button.dart';
 
 const Map<Enum$GridSectionSize, int> _columnSize = {
   Enum$GridSectionSize.half: kIsWeb ? 8 : 2,
