+++ lib/components/sections/icon_section.dart
@@ -1,11 +1,11 @@
-import 'package:brunstadtv_app/components/section_item_click_wrapper.dart';
+import 'package:brunstadtv_app/components/sections/section_item_click_wrapper.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import '../models/analytics/sections.dart';
-import 'category_button.dart';
-import 'horizontal_slider.dart';
-import '../graphql/queries/page.graphql.dart';
+import '../../models/analytics/sections.dart';
+import '../category_button.dart';
+import '../horizontal_slider.dart';
+import '../../graphql/queries/page.graphql.dart';
 
 class IconSection extends StatelessWidget {
   final Fragment$Section$$IconSection data;
