+++ lib/components/sections/page_section.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import '../graphql/queries/page.graphql.dart';
-import '../theme/design_system/design_system.dart';
+import '../../graphql/queries/page.graphql.dart';
+import '../../theme/design_system/design_system.dart';
 
 class PageSection extends StatelessWidget {
   final String? title;
