+++ lib/components/sections/section_update_handler.dart
@@ -3,12 +3,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../graphql/client.dart';
-import '../graphql/queries/page.graphql.dart';
-import '../helpers/event_bus.dart';
-import '../models/events/my_list_changed.dart';
-import '../providers/section_updates.dart';
-import '../helpers/extensions.dart';
+import '../../graphql/client.dart';
+import '../../graphql/queries/page.graphql.dart';
+import '../../helpers/event_bus.dart';
+import '../../models/events/my_list_changed.dart';
+import '../../providers/section_updates.dart';
+import '../../helpers/extensions.dart';
 import 'section.dart';
 
 class SectionUpdateHandler extends HookConsumerWidget {
